refactor(neo4j): decompose monolithic neo4j_graph.py into modular package
This PR decomposes the monolithic `src/waggle/neo4j_graph.py` file into a modular package structure under `src/waggle/neo4j/`.

- Extracted traversal, mutation, base, and transcript logic into dedicated mixins/classes under the `waggle/neo4j` package.
- Solved the indentation issue for the 47 trapped methods.
- Replaced the monolithic `neo4j_graph.py` file with a wrapper facade importing from `waggle.neo4j` for backwards compatibility.
- Fixed schema version and type checking imports.
- Standardized methods to align with SQLite backend API signatures.

Closes #363
